Correccion renderizado doble, y mostrar contador en segundos

+++ app/Livewire/DynamicCode.php
@@ -8,21 +8,26 @@ use Livewire\Component;
 class DynamicCode extends Component
 {
     public $code;
+
+    protected $doubleRenderLimit = false;
     protected $lastPassword;
 
-    // 1 = 0.25seg
+    // 1 = 0.1seg
+    public $toSecondMultiplier = 4;
     public $timeLeft;
     public $deleteTimeLeft = -1;
-    public $maxTime = 160;
+    public $durationSeconds;
 
-    public function mount()
+
+    public function mount($durationSeconds)
     {
+        $this->durationSeconds = $durationSeconds * $this->toSecondMultiplier;
         $password = auth()->user()->active_dynamic_password();
         if ($password){
             $this->lastPassword = $password;
             $this->code = $password->code;
             $this->code = substr_replace($this->code, '-', 3, 0);
-            $this->timeLeft = now()->diffInSeconds($this->lastPassword->expiration_time) * 4;
+            $this->timeLeft = now()->diffInSeconds($this->lastPassword->expiration_time) * $this->toSecondMultiplier;
         }
         else{
             DynamicPassword::where('user_id', auth()->user()->id)->delete();
@@ -37,22 +42,26 @@ class DynamicCode extends Component
             [
                 'code' => $this->code,
                 'user_id' => auth()->user()->id,
-                'expiration_time' => now()->addSeconds($this->maxTime / 4),
+                'expiration_time' => now()->addSeconds($this->durationSeconds / $this->toSecondMultiplier),
             ]);
         $this->code = substr_replace($this->code, '-', 3, 0);
-        $this->timeLeft = $this->maxTime;
+        $this->timeLeft = $this->durationSeconds;
     }
 
     public function decrementTimeLeft()
     {
-        $this->timeLeft--;
-        $this->deleteTimeLeft = $this->deleteTimeLeft === -1 ? $this->timeLeft + 60 : --$this->deleteTimeLeft;
-        if ($this->timeLeft < -2) {
-            $this->generateCode();
-        }
-        if ($this->deleteTimeLeft < 0) {
-            DynamicPassword::firstWhere('user_id', auth()->user()->id)->delete();
-            $this->deleteTimeLeft = $this->timeLeft + 60;
+        $this->doubleRenderLimit = !$this->doubleRenderLimit;
+        if (!$this->doubleRenderLimit){
+            $this->timeLeft--;
+            $secondsToDelete = 15 * $this->toSecondMultiplier;
+            $this->deleteTimeLeft = $this->deleteTimeLeft === -1 ? $this->timeLeft + $secondsToDelete : --$this->deleteTimeLeft;
+            if ($this->timeLeft < 0) {
+                $this->generateCode();
+            }
+            if ($this->deleteTimeLeft < 0) {
+                DynamicPassword::firstWhere('user_id', auth()->user()->id)->delete();
+                $this->deleteTimeLeft = $this->timeLeft + $secondsToDelete;
+            }
         }
     }
